pyproject.toml
[tool.poetry]
name = "lesson-57"
version = "0.1.0"
description = ""
authors = ["Saken Myrzabekov <myrzabekovsaken89@gmail.com>"]
readme = "README.md"
packages = [{include = "lesson_57"}]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django = "^4.2.4"
markdown = "^3.4.4"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
